+++ core/system_env.py
@@ -1,6 +1,6 @@
 # ==============================================================================
-# RTMS — Real-Time Multicam System
-# Desarrollado y soporte: Joaquín Yarsky - joaquinyarsky@gmail.com
+# RTMS v2.1.0 — Real-Time Multicam System
+# Desarrollado y soporte: Joaquín Yarsky - joaquinyarsky@gmail.com - +54 2625-437980
 # ==============================================================================
 
 import subprocess